+++ app/Http/Requests/InspectTrackingProviderMixTokenRequest.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace App\Http\Requests;
+
+class InspectTrackingProviderMixTokenRequest extends BaseRequest
+{
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    public function rules(): array
+    {
+        return [
+            'merchant_id' => ['required', 'uuid', 'exists:merchants,uuid'],
+        ];
+    }
+}